+++ jeap-process-archive-type-registry-maven-plugin/src/main/java/ch/admin/bit/jeap/processarchive/registry/verifier/FileNotChangedValidator.java
@@ -15,7 +15,7 @@ public class FileNotChangedValidator {
         File oldDescriptorFolder = descriptorDir(validationContext, true);
 
         return Arrays.stream(Objects.requireNonNullElse(oldDescriptorFolder.list(), new String[0]))
-                .map(filename -> fileNotDeleted(newDescriptorFolder, filename))
+                .map(filename -> fileNotDeleted(newDescriptorFolder, filename, validationContext.getTrunkBranchName()))
                 .reduce(ValidationResult.ok(), ValidationResult::merge);
     }
 
@@ -29,7 +29,7 @@ public class FileNotChangedValidator {
         }
 
         return Arrays.stream(Objects.requireNonNullElse(oldSystemDir.list(), new String[0]))
-                .map(filename -> fileNotDeleted(newSystemDir, filename))
+                .map(filename -> fileNotDeleted(newSystemDir, filename, validationContext.getTrunkBranchName()))
                 .reduce(ValidationResult.ok(), ValidationResult::merge);
     }
 
@@ -43,21 +43,21 @@ public class FileNotChangedValidator {
         }
 
         if (!oldCommonRootDir.exists()) {
-            String message = "Common dir does not exist in master but does in this branch. You are not allowed to add a common dir on root level";
+            String message = String.format("Common dir does not exist in %s but does in this branch. You are not allowed to add a common dir on root level", validationContext.getTrunkBranchName());
             return ValidationResult.fail(message);
         }
 
         if (!newCommonRootDir.exists()) {
-            String message = "Common dir does exist in master but does not in this branch. You are not allowed to delete common dir on root level";
+            String message = String.format("Common dir does exist in %s but does not in this branch. You are not allowed to delete common dir on root level", validationContext.getTrunkBranchName());
             return ValidationResult.fail(message);
         }
 
         ValidationResult filesCreated = Arrays.stream(Objects.requireNonNullElse(newCommonRootDir.list(), new String[0]))
-                .map(filename -> fileExistedBefore(oldCommonRootDir, newCommonRootDir, filename))
+                .map(filename -> fileExistedBefore(oldCommonRootDir, newCommonRootDir, filename, validationContext.getTrunkBranchName()))
                 .reduce(ValidationResult.ok(), ValidationResult::merge);
 
         ValidationResult filesChanged = Arrays.stream(Objects.requireNonNullElse(oldCommonRootDir.list(), new String[0]))
-                .map(filename -> fileNotChanged(oldCommonRootDir, newCommonRootDir, filename))
+                .map(filename -> fileNotChanged(oldCommonRootDir, newCommonRootDir, filename, validationContext.getTrunkBranchName()))
                 .reduce(ValidationResult.ok(), ValidationResult::merge);
 
         return ValidationResult.merge(filesCreated, filesChanged);
@@ -74,14 +74,14 @@ public class FileNotChangedValidator {
 
         if (!newCommonSystemDir.exists()) {
             String message = String.format("""
-                    Common dir for system %s exist in master but does in this branch.\
+                    Common dir for system %s exist in %s but does in this branch.\
                      You are not allowed to delete a common dir\
-                    """, validationContext.getSystemName());
+                    """, validationContext.getSystemName(), validationContext.getTrunkBranchName());
             return ValidationResult.fail(message);
         }
 
         return Arrays.stream(Objects.requireNonNullElse(oldCommonSystemDir.list(), new String[0]))
-                .map(filename -> fileNotChanged(oldCommonSystemDir, newCommonSystemDir, filename))
+                .map(filename -> fileNotChanged(oldCommonSystemDir, newCommonSystemDir, filename, validationContext.getTrunkBranchName()))
                 .reduce(ValidationResult.ok(), ValidationResult::merge);
     }
 
@@ -97,40 +97,40 @@ public class FileNotChangedValidator {
         return Arrays.stream(Objects.requireNonNullElse(oldArchiveTypeDir.list(), new String[0]))
                 //Archive type descriptors can change
                 .filter(filename -> !filename.endsWith("json"))
-                .map(filename -> fileNotChanged(oldArchiveTypeDir, newArchiveTypeDir, filename))
+                .map(filename -> fileNotChanged(oldArchiveTypeDir, newArchiveTypeDir, filename, validationContext.getTrunkBranchName()))
                 .reduce(ValidationResult.ok(), ValidationResult::merge);
     }
 
-    private static ValidationResult fileExistedBefore(File oldFolder, File newFolder, String filename) {
+    private static ValidationResult fileExistedBefore(File oldFolder, File newFolder, String filename, String trunkBranchName) {
         File newFile = new File(newFolder, filename);
         File oldFile = new File(oldFolder, filename);
         if (!oldFile.exists()) {
             String message = String.format("""
-                    File %s does not exist master but does  exist in this branch. \
+                    File %s does not exist %s but does  exist in this branch. \
                     You are not allowed to create this file\
-                    """, newFile.getAbsolutePath());
+                    """, newFile.getAbsolutePath(), trunkBranchName);
             return ValidationResult.fail(message);
         }
         return ValidationResult.ok();
     }
 
-    private static ValidationResult fileNotChanged(File oldFolder, File newFolder, String filename) {
+    private static ValidationResult fileNotChanged(File oldFolder, File newFolder, String filename, String trunkBranchName) {
         File newFile = new File(newFolder, filename);
         File oldFile = new File(oldFolder, filename);
 
         if (!newFile.exists()) {
             String message = String.format("""
-                    File %s exists in master but does not exist in this branch. \
+                    File %s exists in %s but does not exist in this branch. \
                     You are not allowed to delete this file\
-                    """, newFile.getAbsolutePath());
+                    """, newFile.getAbsolutePath(), trunkBranchName);
             return ValidationResult.fail(message);
         }
         try {
             if (FileUtils.checksumCRC32(newFile) != FileUtils.checksumCRC32(oldFile)) {
                 String message = String.format("""
-                        File %s has changed compared to master. \
+                        File %s has changed compared to %s. \
                         You are not allowed to change this file\
-                        """, newFile.getAbsolutePath());
+                        """, newFile.getAbsolutePath(), trunkBranchName);
                 return ValidationResult.fail(message);
             }
         } catch (IOException e) {
@@ -140,14 +140,14 @@ public class FileNotChangedValidator {
         return ValidationResult.ok();
     }
 
-    private static ValidationResult fileNotDeleted(File newFolder, String filename) {
+    private static ValidationResult fileNotDeleted(File newFolder, String filename, String trunkBranchName) {
         File newFile = new File(newFolder, filename);
 
         if (!newFile.exists()) {
             String message = String.format("""
-                    File %s exists in master but does not exist in this branch. \
+                    File %s exists in %s but does not exist in this branch. \
                     You are not allowed to delete this file\
-                    """, newFile.getAbsolutePath());
+                    """, newFile.getAbsolutePath(), trunkBranchName);
             return ValidationResult.fail(message);
         }
         return ValidationResult.ok();