+++ jeap-process-archive-type-registry-maven-plugin/src/main/java/ch/admin/bit/jeap/processarchive/registry/verifier/ValidationContext.java
@@ -17,4 +17,5 @@ public class ValidationContext {
     File descriptor;
     String archiveTypeName;
     File archiveTypeDir;
+    String trunkBranchName;
 }