+++ jeap-process-archive-type-registry-maven-plugin/src/test/java/ch/admin/bit/jeap/processarchive/registry/verifier/FileNotChangedValidatorTest.java
@@ -35,6 +35,7 @@ class FileNotChangedValidatorTest {
                 .descriptorDir(newDescriptor)
                 .systemName("test")
                 .archiveTypeName("decree")
+                .trunkBranchName("master")
                 .build();
         oldCommon = new File(oldDescriptor, ArchiveTypeRegistryConstants.COMMON_DIR_NAME);
         newCommon = new File(newDescriptor, ArchiveTypeRegistryConstants.COMMON_DIR_NAME);