+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [16.2.0] - 2026-07-07
+
+### Added
+
+- Made trunk branch name configurable in jeap-process-archive-type-registry-maven-plugin 
+
 ## [16.1.0] - 2026-07-06
 
 ### Changed